+++ gentle/__main__.py
@@ -14,6 +14,7 @@ import gentle
 from gentle.generators import AbstractGenerator, GeneratorClass
 from gentle.metadata import MetadataXML
 
+import gentle.generators.autoconf
 import gentle.generators.bower
 import gentle.generators.cargo
 import gentle.generators.composer