+++ tests/conftest.py
@@ -13,20 +13,27 @@ from gentle.pms.portagepm import parse_mxml
 def pytest_addoption(parser):
     parser.addoption("--net", action="store_true",
                      help="run tests that connect to network")
+    parser.addoption("--with-perl", action="store_true",
+                     help="run tests that require Perl")
 
 
 def pytest_configure(config):
     config.addinivalue_line("markers", "net: mark test as using network")
+    config.addinivalue_line("markers", "perl: mark test as using Perl")
 
 
 def pytest_collection_modifyitems(config, items):
-    if config.getoption("--net"):
-        # --net given in cli: do not skip network tests
-        return
-    skip_net = pytest.mark.skip(reason="need --net option to run")
-    for item in items:
-        if "net" in item.keywords:
-            item.add_marker(skip_net)
+    if not config.getoption("--net"):
+        skip_net = pytest.mark.skip(reason="need --net option to run")
+        for item in items:
+            if "net" in item.keywords:
+                item.add_marker(skip_net)
+
+    if not config.getoption("--with-perl"):
+        skip_perl = pytest.mark.skip(reason="need --with-perl option to run")
+        for item in items:
+            if "perl" in item.keywords:
+                item.add_marker(skip_perl)
 
 
 @pytest.fixture