+++ docs/release-notes.rst
@@ -7,8 +7,9 @@ Release Notes
 
 0.4.0
 -----
-* New generator: PEAR/PECL
 * New generator: Apache Maven POM
+* New generator: Dart Pubspec
+* New generator: PEAR/PECL
 * Add ``kde-invent`` remote-id
 * Trim ".git" suffix when extracting remote-id